better input
previously it would delete plaintext, might not want that. So now, instead of like an actual queue where you remove elements, this time instead you maintain a variable of where you have calculated till. Lastly if ur position var is ever greater than the length of the message, that means the user deleted plaintext and typed something else in which case set pos to 0. IMPORTANT: you can't mistype because it will use the wrong rotor positions and decryption won't work